fix: custom embedded fields for employee entity
